@@ -25,16 +25,20 @@
         "classmap" : ["lib/"]
     },
     "require": {
-        "php": ">=5.3.0",
-        "ext-gd": "*",
+        "php": ">=5.4.0",
         "ext-dom": "*",
         "ext-mbstring": "*",
         "phenx/php-font-lib": "0.5.*",
-        "phenx/php-svg-lib": "0.2.*"
+        "phenx/php-svg-lib": "0.3.*"
     },
     "require-dev": {
-        "phpunit/phpunit": "4.8.*",
+        "phpunit/phpunit": "^4.8|^5.5|^6.5",
         "squizlabs/php_codesniffer": "2.*"
+    },
+    "suggest": {
+        "ext-gd": "Needed to process images",
+        "ext-imagick": "Improves image processing performance",
+        "ext-gmagick": "Improves image processing performance"
     },
     "extra": {
         "branch-alias": {
