@@ -7,18 +7,29 @@
  * @license http://www.gnu.org/copyleft/lesser.html GNU Lesser General Public License
  */
 
-/**
- * Dompdf autoload function
- *
- * If you have an existing autoload function, add a call to this function
- * from your existing __autoload() implementation.
- *
- * @param string $class
- */
-
+// HMLT5 Parser
 require_once __DIR__ . '/lib/html5lib/Parser.php';
+
+// Sabberworm
+spl_autoload_register(function($class)
+{
+    if (strpos($class, 'Sabberworm') !== false) {
+        $file = str_replace('\\', DIRECTORY_SEPARATOR, $class);
+        $file = realpath(__DIR__ . '/lib/php-css-parser/lib/' . (empty($file) ? '' : DIRECTORY_SEPARATOR) . $file . '.php');
+        if (file_exists($file)) {
+            require_once $file;
+            return true;
+        }
+    }
+    return false;
+});
+
+// php-font-lib
 require_once __DIR__ . '/lib/php-font-lib/src/FontLib/Autoloader.php';
+
+//php-svg-lib
 require_once __DIR__ . '/lib/php-svg-lib/src/autoload.php';
+
 
 /*
  * New PHP 5.3.0 namespaced autoloader
