@@ -92,5 +92,4 @@ return array(
             'italic' => $distFontDir . 'DejaVuSerif-Italic',
             'normal' => $distFontDir . 'DejaVuSerif'
         )
-)
-?>
+);
