@@ -2,8 +2,8 @@
 /**
  * @package php-svg-lib
  * @link    http://github.com/PhenX/php-svg-lib
- * @author  Fabien M�nager <fabien.menager@gmail.com>
- * @license http://www.gnu.org/copyleft/lesser.html GNU Lesser General Public License
+ * @author  Fabien Ménager <fabien.menager@gmail.com>
+ * @license GNU LGPLv3+ http://www.gnu.org/copyleft/lesser.html
  */
 
 namespace Svg;
@@ -333,6 +333,9 @@ class Document extends AbstractTag
             case 'text':
                 $tag = new Text($this, $name);
                 break;
+
+            case 'desc':
+                return;
         }
 
         if ($tag) {
@@ -350,8 +353,6 @@ class Document extends AbstractTag
             $this->stack[] = $tag;
 
             $tag->handle($attributes);
-        } else {
-            echo "Unknown: '$name'\n";
         }
     }
 
