@@ -2,8 +2,8 @@
 /**
  * @package php-svg-lib
  * @link    http://github.com/PhenX/php-svg-lib
- * @author  Fabien M�nager <fabien.menager@gmail.com>
- * @license http://www.gnu.org/copyleft/lesser.html GNU Lesser General Public License
+ * @author  Fabien Ménager <fabien.menager@gmail.com>
+ * @license GNU LGPLv3+ http://www.gnu.org/copyleft/lesser.html
  */
 
 namespace Svg\Tag;
@@ -33,8 +33,8 @@ class Shape extends AbstractTag
         if ($this->hasShape) {
             $style = $surface->getStyle();
 
-            $fill   = $style->fill   && $style->fill   !== "none";
-            $stroke = $style->stroke && $style->stroke !== "none";
+            $fill   = $style->fill   && is_array($style->fill);
+            $stroke = $style->stroke && is_array($style->stroke);
 
             if ($fill) {
                 if ($stroke) {
