@@ -89,6 +89,7 @@ class CPDF implements Canvas
         "sra3" => array(0, 0, 907.09, 1275.59),
         "sra4" => array(0, 0, 637.80, 907.09),
         "letter" => array(0, 0, 612.00, 792.00),
+        "half-letter" => array(0, 0, 396.00, 612.00),
         "legal" => array(0, 0, 612.00, 1008.00),
         "ledger" => array(0, 0, 1224.00, 792.00),
         "tabloid" => array(0, 0, 792.00, 1224.00),
@@ -151,7 +152,7 @@ class CPDF implements Canvas
     private $_page_text;
 
     /**
-     * Array of pages for accesing after rendering is initially complete
+     * Array of pages for accessing after rendering is initially complete
      *
      * @var array
      */
@@ -241,8 +242,12 @@ class CPDF implements Canvas
                 continue;
             }
 
-            if ($this->_dompdf->getOptions()->getDebugPng()) print '[__destruct unlink ' . $img . ']';
-            if (!$this->_dompdf->getOptions()->getDebugKeepTemp()) unlink($img);
+            if ($this->_dompdf->getOptions()->getDebugPng()) {
+                print '[__destruct unlink ' . $img . ']';
+            }
+            if (!$this->_dompdf->getOptions()->getDebugKeepTemp()) {
+                unlink($img);
+            }
         }
     }
 
@@ -270,14 +275,14 @@ class CPDF implements Canvas
     /**
      * Opens a new 'object'
      *
-     * While an object is open, all drawing actions are recored in the object,
+     * While an object is open, all drawing actions are recorded in the object,
      * as opposed to being drawn on the current page.  Objects can be added
      * later to a specific page or to several pages.
      *
      * The return value is an integer ID for the new object.
      *
-     * @see CPDF_Adapter::close_object()
-     * @see CPDF_Adapter::add_object()
+     * @see CPDF::close_object()
+     * @see CPDF::add_object()
      *
      * @return int
      */
@@ -291,7 +296,7 @@ class CPDF implements Canvas
     /**
      * Reopens an existing 'object'
      *
-     * @see CPDF_Adapter::open_object()
+     * @see CPDF::open_object()
      * @param int $object the ID of a previously opened object
      */
     public function reopen_object($object)
@@ -303,7 +308,7 @@ class CPDF implements Canvas
     /**
      * Closes the current 'object'
      *
-     * @see CPDF_Adapter::open_object()
+     * @see CPDF::open_object()
      */
     public function close_object()
     {
@@ -553,6 +558,26 @@ class CPDF implements Canvas
 
         $this->_pdf->line($x1, $this->y($y1),
             $x2, $this->y($y2));
+        $this->_set_line_transparency("Normal", $this->_current_opacity);
+    }
+
+    /**
+     * Draw line at the specified coordinates on every page.
+     *
+     * See {@link Style::munge_color()} for the format of the colour array.
+     *
+     * @param float $x1
+     * @param float $y1
+     * @param float $x2
+     * @param float $y2
+     * @param array $color
+     * @param float $width
+     * @param array $style optional
+     */
+    public function page_line($x1, $y1, $x2, $y2, $color, $width, $style = array())
+    {
+        $_t = 'line';
+        $this->_page_text[] = compact('_t', 'x1', 'y1', 'x2', 'y2', 'color', 'width', 'style');
     }
 
     /**
@@ -572,6 +597,7 @@ class CPDF implements Canvas
         $this->_set_line_style($width, "butt", "", $style);
 
         $this->_pdf->ellipse($x, $this->y($y), $r1, $r2, 0, 8, $astart, $aend, false, false, true, false);
+        $this->_set_line_transparency("Normal", $this->_current_opacity);
     }
 
     /**
@@ -601,7 +627,7 @@ class CPDF implements Canvas
             imageinterlace($im, false);
 
             $tmp_dir = $this->_dompdf->getOptions()->getTempDir();
-            $tmp_name = tempnam($tmp_dir, "{$type}dompdf_img_");
+            $tmp_name = @tempnam($tmp_dir, "{$type}dompdf_img_");
             @unlink($tmp_name);
             $filename = "$tmp_name.png";
             $this->_image_cache[] = $filename;
@@ -631,6 +657,7 @@ class CPDF implements Canvas
         $this->_set_stroke_color($color);
         $this->_set_line_style($width, "butt", "", $style);
         $this->_pdf->rectangle($x1, $this->y($y1) - $h, $w, $h);
+        $this->_set_line_transparency("Normal", $this->_current_opacity);
     }
 
     /**
@@ -644,6 +671,7 @@ class CPDF implements Canvas
     {
         $this->_set_fill_color($color);
         $this->_pdf->filledRectangle($x1, $this->y($y1) - $h, $w, $h);
+        $this->_set_fill_transparency("Normal", $this->_current_opacity);
     }
 
     /**
@@ -768,6 +796,9 @@ class CPDF implements Canvas
         }
 
         $this->_pdf->polygon($points, count($points) / 2, $fill);
+
+        $this->_set_fill_transparency("Normal", $this->_current_opacity);
+        $this->_set_line_transparency("Normal", $this->_current_opacity);
     }
 
     /**
@@ -789,6 +820,9 @@ class CPDF implements Canvas
         }
 
         $this->_pdf->ellipse($x, $this->y($y), $r1, 0, 0, 8, 0, 360, 1, $fill);
+
+        $this->_set_fill_transparency("Normal", $this->_current_opacity);
+        $this->_set_line_transparency("Normal", $this->_current_opacity);
     }
 
     /**
@@ -811,7 +845,9 @@ class CPDF implements Canvas
 
         switch ($type) {
             case "jpeg":
-                if ($debug_png) print '!!!jpg!!!';
+                if ($debug_png) {
+                    print '!!!jpg!!!';
+                }
                 $this->_pdf->addJpegFromFile($img, $x, $this->y($y) - $h, $w, $h);
                 break;
 
@@ -889,6 +925,8 @@ class CPDF implements Canvas
         //
         //$pdf->addText($x, $this->y($y) - ($pdf->fonts[$pdf->currentFont]['FontBBox'][3]*$size)/1000, $size, $text, $angle, $word_space, $char_space);
         $pdf->addText($x, $this->y($y) - $pdf->getFontHeight($size), $size, $text, $angle, $word_space, $char_space);
+
+        $this->_set_fill_transparency("Normal", $this->_current_opacity);
     }
 
     /**
@@ -1073,6 +1111,10 @@ class CPDF implements Canvas
                         }
                         $eval->evaluate($code, array('PAGE_NUM' => $page_number, 'PAGE_COUNT' => $this->_page_count));
                         break;
+
+                    case 'line':
+                        $this->line( $x1, $y1, $x2, $y2, $color, $width, $style );
+                        break;
                 }
             }
 
@@ -1082,31 +1124,50 @@ class CPDF implements Canvas
     }
 
     /**
-     * Streams the PDF directly to the browser
+     * Streams the PDF to the client.
      *
-     * @param string $filename the name of the PDF file
-     * @param array $options associative array, 'Attachment' => 0 or 1, 'compress' => 1 or 0
+     * @param string $filename The filename to present to the client.
+     * @param array $options Associative array: 'compress' => 1 or 0 (default 1); 'Attachment' => 1 or 0 (default 1).
      */
-    public function stream($filename, $options = null)
+    public function stream($filename = "document.pdf", $options = array())
     {
-        // Add page text
+        if (headers_sent()) {
+            die("Unable to stream pdf: headers already sent");
+        }
+
+        if (!isset($options["compress"])) $options["compress"] = true;
+        if (!isset($options["Attachment"])) $options["Attachment"] = true;
+
         $this->_add_page_text();
 
-        $options["Content-Disposition"] = $filename;
-        $this->_pdf->stream($options);
+        $debug = !$options['compress'];
+        $tmp = ltrim($this->_pdf->output($debug));
+
+        header("Cache-Control: private");
+        header("Content-Type: application/pdf");
+        header("Content-Length: " . mb_strlen($tmp, "8bit"));
+
+        $filename = str_replace(array("\n", "'"), "", basename($filename, ".pdf")) . ".pdf";
+        $attachment = $options["Attachment"] ? "attachment" : "inline";
+        header(Helpers::buildContentDispositionHeader($attachment, $filename));
+
+        echo $tmp;
+        flush();
     }
 
     /**
-     * Returns the PDF as a string
+     * Returns the PDF as a string.
      *
-     * @param array $options Output options
+     * @param array $options Associative array: 'compress' => 1 or 0 (default 1).
      * @return string
      */
-    public function output($options = null)
+    public function output($options = array())
     {
+        if (!isset($options["compress"])) $options["compress"] = true;
+
         $this->_add_page_text();
 
-        $debug = isset($options["compress"]) && $options["compress"] != 1;
+        $debug = !$options['compress'];
 
         return $this->_pdf->output($debug);
     }
