@@ -107,7 +107,7 @@ class Dompdf
     /**
      * Desired paper size ('letter', 'legal', 'A4', etc.)
      *
-     * @var string
+     * @var string|array
      */
     private $paperSize;
 
@@ -273,6 +273,11 @@ class Dompdf
     {
         mb_internal_encoding('UTF-8');
 
+        if (version_compare(PHP_VERSION, '7.0.0') >= 0)
+        {
+            ini_set('pcre.jit', 0);
+        }
+
         if (isset($options) && $options instanceof Options) {
             $this->setOptions($options);
         } elseif (is_array($options)) {
@@ -368,7 +373,7 @@ class Dompdf
 
             $ext = strtolower(pathinfo($realfile, PATHINFO_EXTENSION));
             if (!in_array($ext, $this->allowedLocalFileExtensions)) {
-                throw new Exception("Permission denied on $file.");
+                throw new Exception("Permission denied on $file. This file extension is forbidden");
             }
 
             if (!$realfile) {
@@ -397,8 +402,8 @@ class Dompdf
     }
 
     /**
-     * @param $str
-     * @param null $encoding
+     * @param string $str
+     * @param string $encoding
      * @deprecated
      */
     public function load_html($str, $encoding = 'UTF-8')
@@ -465,7 +470,7 @@ class Dompdf
         // https://developer.mozilla.org/en/mozilla's_quirks_mode
         $quirksmode = false;
 
-        if ($this->options->isHtml5ParserEnabled()) {
+        if ($this->options->isHtml5ParserEnabled() && class_exists("HTML5_Tokenizer")) {
             $tokenizer = new HTML5_Tokenizer($str);
             $tokenizer->parse();
             $doc = $tokenizer->save();
@@ -489,6 +494,16 @@ class Dompdf
             $doc->preserveWhiteSpace = true;
             $doc->loadHTML($str);
             $doc->encoding = $encoding;
+
+            // Remove #text children nodes in nodes that shouldn't have
+            $tag_names = array("html", "table", "tbody", "thead", "tfoot", "tr");
+            foreach ($tag_names as $tag_name) {
+                $nodes = $doc->getElementsByTagName($tag_name);
+
+                foreach ($nodes as $node) {
+                    self::removeTextNodes($node);
+                }
+            }
 
             // If some text is before the doctype, we are in quirksmode
             if (preg_match("/^(.+)<!doctype/i", ltrim($str), $matches)) {
@@ -597,7 +612,7 @@ class Dompdf
                             if (!$accept) {
                                 //found at least one mediatype, but none of the accepted ones
                                 //Skip this css file.
-                                continue;
+                                break;
                             }
                         }
 
@@ -618,7 +633,7 @@ class Dompdf
                         ($media = $tag->getAttribute("media")) &&
                         !in_array($media, $acceptedmedia)
                     ) {
-                        continue;
+                        break;
                     }
 
                     $css = "";
@@ -755,6 +770,7 @@ class Dompdf
                 if ($style->display === "list-item") {
                     $chars = ListBullet::get_counter_chars($style->list_style_type);
                     $canvas->register_string_subset($style->font_family, $chars);
+                    $canvas->register_string_subset($style->font_family, '.');
                     continue;
                 }
 
@@ -897,15 +913,10 @@ class Dompdf
     }
 
     /**
-     * Streams the PDF to the client
+     * Streams the PDF to the client.
      *
      * The file will open a download dialog by default. The options
      * parameter controls the output. Accepted options (array keys) are:
-     *
-     * 'Accept-Ranges' => 1 or 0 (=default): Send an 'Accept-Ranges:'
-     *   HTTP header, see https://tools.ietf.org/html/rfc2616#section-14.5
-     *   This header seems to have caused some problems, despite the fact
-     *   that it is supposed to solve them, so I am leaving it off by default.
      *
      * 'compress' = > 1 (=default) or 0:
      *   Apply content stream compression
@@ -917,7 +928,7 @@ class Dompdf
      * @param string $filename the name of the streamed file
      * @param array $options header options (see above)
      */
-    public function stream($filename = 'document.pdf', $options = null)
+    public function stream($filename = "document.pdf", $options = array())
     {
         $this->saveLocale();
 
@@ -930,21 +941,18 @@ class Dompdf
     }
 
     /**
-     * Returns the PDF as a string
+     * Returns the PDF as a string.
      *
-     * The file will open a download dialog by default.  The options
-     * parameter controls the output.  Accepted options are:
-     *
+     * The options parameter controls the output. Accepted options are:
      *
      * 'compress' = > 1 or 0 - apply content stream compression, this is
      *    on (1) by default
      *
-     *
      * @param array $options options (see above)
      *
-     * @return string
+     * @return string|null
      */
-    public function output($options = null)
+    public function output($options = array())
     {
         $this->saveLocale();
 
@@ -1027,7 +1035,7 @@ class Dompdf
     /**
      * Sets the paper size & orientation
      *
-     * @param string $size 'letter', 'legal', 'A4', etc. {@link Dompdf\Adapter\CPDF::$PAPER_SIZES}
+     * @param string|array $size 'letter', 'legal', 'A4', etc. {@link Dompdf\Adapter\CPDF::$PAPER_SIZES}
      * @param string $orientation 'portrait' or 'landscape'
      * @return $this
      */
@@ -1042,7 +1050,7 @@ class Dompdf
      * Gets the paper size
      *
      * @param null|string|array $paperSize
-     * @return \int[] A four-element integer array
+     * @return int[] A four-element integer array
      */
     public function getPaperSize($paperSize = null)
     {
