@@ -84,10 +84,10 @@ class FontMetrics
      * Saves the stored font family cache
      *
      * The name and location of the cache file are determined by {@link
-     * FontMetrics::CACHE_FILE}.  This file should be writable by the
+     * FontMetrics::CACHE_FILE}. This file should be writable by the
      * webserver process.
      *
-     * @see Font_Metrics::load_font_families()
+     * @see FontMetrics::loadFontFamilies()
      */
     public function saveFontFamilies()
     {
@@ -118,14 +118,14 @@ class FontMetrics
     /**
      * Loads the stored font family cache
      *
-     * @see save_font_families()
+     * @see FontMetrics::saveFontFamilies()
      */
     public function loadFontFamilies()
     {
         $fontDir = $this->getOptions()->getFontDir();
         $rootDir = $this->getOptions()->getRootDir();
 
-        // FIXME: tempoarary define constants for cache files <= v0.6.2
+        // FIXME: temporarily define constants for cache files <= v0.6.2
         if (!defined("DOMPDF_DIR")) { define("DOMPDF_DIR", $rootDir); }
         if (!defined("DOMPDF_FONT_DIR")) { define("DOMPDF_FONT_DIR", $fontDir); }
 
@@ -170,7 +170,6 @@ class FontMetrics
      */
     public function registerFont($style, $remoteFile, $context = null)
     {
-        $fontDir = $this->getOptions()->getFontDir();
         $fontname = mb_strtolower($style["family"]);
         $families = $this->getFontFamilies();
 
@@ -179,51 +178,56 @@ class FontMetrics
             $entry = $families[$fontname];
         }
 
-        $localFile = $fontDir . DIRECTORY_SEPARATOR . md5($remoteFile);
-        $localTempFile = $this->options->get('tempDir') . "/" . md5($remoteFile);
-        $cacheEntry = $localFile;
-        $localFile .= ".".strtolower(pathinfo($remoteFile,PATHINFO_EXTENSION));
-
         $styleString = $this->getType("{$style['weight']} {$style['style']}");
-
-        if ( !isset($entry[$styleString]) ) {
-            $entry[$styleString] = $cacheEntry;
-
-            // Download the remote file
-            list($remoteFileContent, $http_response_header) = @Helpers::getFileContent($remoteFile, $context);
-            if (false === $remoteFileContent) {
-                return false;
-            }
-            file_put_contents($localTempFile, $remoteFileContent);
-
-            $font = Font::load($localTempFile);
-
-            if (!$font) {
-                unlink($localTempFile);
-                return false;
-            }
-
-            $font->parse();
-            $font->saveAdobeFontMetrics("$cacheEntry.ufm");
-            $font->close();
-
-            unlink($localTempFile);
-
-            if ( !file_exists("$cacheEntry.ufm") ) {
-                return false;
-            }
-
-            // Save the changes
-            file_put_contents($localFile, $remoteFileContent);
-
-            if ( !file_exists($localFile) ) {
-                unlink("$cacheEntry.ufm");
-                return false;
-            }
-
-            $this->setFontFamily($fontname, $entry);
-            $this->saveFontFamilies();
+        if (isset($entry[$styleString])) {
+            return true;
         }
+
+        $fontDir = $this->getOptions()->getFontDir();
+        $remoteHash = md5($remoteFile);
+        $localFile = $fontDir . DIRECTORY_SEPARATOR . $remoteHash;
+
+        $cacheEntry = $localFile;
+        $localFile .= ".".strtolower(pathinfo(parse_url($remoteFile, PHP_URL_PATH),PATHINFO_EXTENSION));
+
+        $entry[$styleString] = $cacheEntry;
+
+        // Download the remote file
+        list($remoteFileContent, $http_response_header) = @Helpers::getFileContent($remoteFile, $context);
+        if (false === $remoteFileContent) {
+            return false;
+        }
+
+        $localTempFile = @tempnam($this->options->get("tempDir"), "dompdf-font-");
+        file_put_contents($localTempFile, $remoteFileContent);
+
+        $font = Font::load($localTempFile);
+
+        if (!$font) {
+            unlink($localTempFile);
+            return false;
+        }
+
+        $font->parse();
+        $font->saveAdobeFontMetrics("$cacheEntry.ufm");
+        $font->close();
+
+        unlink($localTempFile);
+
+        if ( !file_exists("$cacheEntry.ufm") ) {
+            return false;
+        }
+
+        // Save the changes
+        file_put_contents($localFile, $remoteFileContent);
+
+        if ( !file_exists($localFile) ) {
+            unlink("$cacheEntry.ufm");
+            return false;
+        }
+
+        $this->setFontFamily($fontname, $entry);
+        $this->saveFontFamilies();
 
         return true;
     }
