@@ -148,7 +148,7 @@ class Frame
     protected $_is_cache = array();
 
     /**
-     * Tells wether the frame was already pushed to the next page
+     * Tells whether the frame was already pushed to the next page
      *
      * @var bool
      */
@@ -160,7 +160,7 @@ class Frame
     public $_float_next_line = false;
 
     /**
-     * Tells wether the frame was split
+     * Tells whether the frame was split
      *
      * @var bool
      */
@@ -325,6 +325,14 @@ class Frame
         $this->_style = null;
         unset($this->_style);
         $this->_style = clone $this->_original_style;
+
+        // If this represents a generated node then child nodes represent generated content.
+        // Remove the children since the content will be generated next time this frame is reflowed. 
+        if ($this->_node->nodeName === "dompdf_generated" && $this->_style->content != "normal") {
+            foreach ($this->get_children() as $child) {
+                $this->remove_child($child);
+            }
+        }
     }
 
     /**
