@@ -10,7 +10,6 @@ namespace Dompdf\FrameDecorator;
 
 use Dompdf\Dompdf;
 use Dompdf\Frame;
-use Dompdf\FontMetrics;
 use Dompdf\Image\Cache;
 
 /**
@@ -48,7 +47,9 @@ class Image extends AbstractFrameDecorator
         $url = $frame->get_node()->getAttribute("src");
 
         $debug_png = $dompdf->getOptions()->getDebugPng();
-        if ($debug_png) print '[__construct ' . $url . ']';
+        if ($debug_png) {
+            print '[__construct ' . $url . ']';
+        }
 
         list($this->_image_url, /*$type*/, $this->_image_msg) = Cache::resolve_url(
             $url,
