@@ -132,7 +132,6 @@ class Table extends AbstractFrameDecorator
         if (count($this->_headers) && !in_array($child, $this->_headers, true) &&
             !in_array($child->get_prev_sibling(), $this->_headers, true)
         ) {
-
             $first_header = null;
 
             // Insert copies of the table headers before $child
@@ -194,10 +193,11 @@ class Table extends AbstractFrameDecorator
      */
     public static function find_parent_table(Frame $frame)
     {
-
-        while ($frame = $frame->get_parent())
-            if ($frame->is_table())
+        while ($frame = $frame->get_parent()) {
+            if ($frame->is_table()) {
                 break;
+            }
+        }
 
         return $frame;
     }
