@@ -43,12 +43,26 @@ class TableRow extends AbstractFrameDecorator
         foreach ($this->get_children() as $child) {
             $display = $child->get_style()->display;
 
-            if ($display !== "table-cell")
+            if ($display !== "table-cell") {
                 $erroneous_frames[] = $child;
+            }
         }
 
         //  dump the extra nodes after the table.
-        foreach ($erroneous_frames as $frame)
+        foreach ($erroneous_frames as $frame) {
             $p->move_after($frame);
+        }
+    }
+
+    function split(Frame $child = null, $force_pagebreak = false)
+    {
+        $this->_already_pushed = true;
+        
+        if (is_null($child)) {
+            parent::split();
+            return;
+        }
+
+        parent::split($child, $force_pagebreak);
     }
 }
