@@ -93,9 +93,9 @@ abstract class AbstractFrameReflower
 
         // Collapse vertical margins:
         $n = $frame->get_next_sibling();
-        if ($n && !$n->is_block()) {
+        if ( $n && !$n->is_block() & !$n->is_table() ) {
             while ($n = $n->get_next_sibling()) {
-                if ($n->is_block()) {
+                if ($n->is_block() || $n->is_table()) {
                     break;
                 }
 
@@ -108,17 +108,19 @@ abstract class AbstractFrameReflower
 
         if ($n) {
             $n_style = $n->get_style();
-            $b = max($b, (float)$n_style->length_in_pt($n_style->margin_top, $cb["h"]));
-            $n_style->margin_top = "0pt";
+            $n_t = (float)$n_style->length_in_pt($n_style->margin_top, $cb["h"]);
+
+            $b = $this->_get_collapsed_margin_length($b, $n_t);
             $style->margin_bottom = $b . "pt";
+            $n_style->margin_top = "0pt";
         }
 
         // Collapse our first child's margin, if there is no border or padding
-        if ($style->get_border_top_width() == 0 && $style->length_in_pt($style->padding_top) == 0) {
+        if ($style->border_top_width == 0 && $style->length_in_pt($style->padding_top) == 0) {
             $f = $this->_frame->get_first_child();
-            if ( $f && !$f->is_block() ) {
+            if ( $f && !$f->is_block() && !$f->is_table() ) {
                 while ( $f = $f->get_next_sibling() ) {
-                    if ( $f->is_block() ) {
+                    if ( $f->is_block() || $f->is_table() ) {
                         break;
                     }
 
@@ -129,21 +131,23 @@ abstract class AbstractFrameReflower
                 }
             }
 
-            // Margin are collapsed only between block elements
+            // Margin are collapsed only between block-level boxes
             if ($f) {
                 $f_style = $f->get_style();
-                $t = max($t, (float)$f_style->length_in_pt($f_style->margin_top, $cb["h"]));
+                $f_t = (float)$f_style->length_in_pt($f_style->margin_top, $cb["h"]);
+
+                $t = $this->_get_collapsed_margin_length($t, $f_t);
                 $style->margin_top = $t."pt";
                 $f_style->margin_top = "0pt";
             }
         }
 
         // Collapse our last child's margin, if there is no border or padding
-        if ($style->get_border_bottom_width() == 0 && $style->length_in_pt($style->padding_bottom) == 0) {
+        if ($style->border_bottom_width == 0 && $style->length_in_pt($style->padding_bottom) == 0) {
             $l = $this->_frame->get_last_child();
-            if ($l && !$l->is_block()) {
+            if ( $l && !$l->is_block() && !$l->is_table() ) {
                 while ( $l = $l->get_prev_sibling() ) {
-                    if ( $l->is_block() ) {
+                    if ( $l->is_block() || $l->is_table() ) {
                         break;
                     }
 
@@ -154,14 +158,38 @@ abstract class AbstractFrameReflower
                 }
             }
 
-            // Margin are collapsed only between block elements
+            // Margin are collapsed only between block-level boxes
             if ($l) {
                 $l_style = $l->get_style();
-                $b = max($b, (float)$l_style->length_in_pt($l_style->margin_bottom, $cb["h"]));
+                $l_b = (float)$l_style->length_in_pt($l_style->margin_bottom, $cb["h"]);
+
+                $b = $this->_get_collapsed_margin_length($b, $l_b);
                 $style->margin_bottom = $b."pt";
                 $l_style->margin_bottom = "0pt";
             }
         }
+    }
+
+    /**
+     * Get the combined (collapsed) length of two adjoining margins.
+     * 
+     * See http://www.w3.org/TR/CSS2/box.html#collapsing-margins.
+     * 
+     * @param number $length1
+     * @param number $length2
+     * @return number
+     */
+    private function _get_collapsed_margin_length($length1, $length2)
+    {
+        if ($length1 < 0 && $length2 < 0) {
+            return min($length1, $length2); // min(x, y) = - max(abs(x), abs(y)), if x < 0 && y < 0
+        }
+        
+        if ($length1 < 0 || $length2 < 0) {
+            return $length1 + $length2; // x + y = x - abs(y), if y < 0
+        }
+        
+        return max($length1, $length2);
     }
 
     /**
@@ -209,10 +237,7 @@ abstract class AbstractFrameReflower
         $low = array();
         $high = array();
 
-        for ($iter = $this->_frame->get_children()->getIterator();
-             $iter->valid();
-             $iter->next()) {
-
+        for ($iter = $this->_frame->get_children()->getIterator(); $iter->valid(); $iter->next()) {
             $inline_min = 0;
             $inline_max = 0;
 
@@ -232,8 +257,12 @@ abstract class AbstractFrameReflower
                 $iter->next();
             }
 
-            if ($inline_max > 0) $high[] = $inline_max;
-            if ($inline_min > 0) $low[] = $inline_min;
+            if ($inline_max > 0) {
+                $high[] = $inline_max;
+            }
+            if ($inline_min > 0) {
+                $low[] = $inline_min;
+            }
 
             if ($iter->valid()) {
                 list($low[], $high[]) = $iter->current()->get_min_max_width();
@@ -369,7 +398,7 @@ abstract class AbstractFrameReflower
                     continue;
                 }
 
-                preg_match('/(counters?)(^\()*?\(\s*([^\s,]+)\s*(,\s*["\']?([^"\'\)]+)["\']?\s*(,\s*([^\s)]+)\s*)?)?\)/i', $match[1], $args);
+                preg_match('/(counters?)(^\()*?\(\s*([^\s,]+)\s*(,\s*["\']?([^"\'\)]*)["\']?\s*(,\s*([^\s)]+)\s*)?)?\)/i', $match[1], $args);
                 $counter_id = $args[3];
                 if (strtolower($args[1]) == 'counter') {
                     // counter(name [,style])
